pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "riptide-proxy"
version = "0.10.0"
description = "Tool to manage development environments for web applications using containers - HTTP and WebSocket Reverse Proxy Server"
readme = "README.rst"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Marco Köpcke", email = "hello@capypara.de" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "riptide-lib >= 0.10, < 0.11",
    "tornado >= 6.5",
    "Click >= 8.2",
    "python-prctl >= 1.8.1; sys_platform == 'linux'",
    "certauth >= 1.3"
]

[project.optional-dependencies]
profiling = ["guppy3 >= 3.0.9"]

[project.urls]
Repository = "https://github.com/theCapypara/riptide-proxy"
Documentation = "https://riptide-docs.readthedocs.io"

[project.scripts]
riptide_proxy = "riptide_proxy.__main__:main"

[tool.ruff]
line-length = 120

[tools.ruff.lint]
select = ["W", "E", "F", "ARG"]

[tool.mypy]
warn_unused_configs = true
mypy_path = "."
explicit_package_bases = true
namespace_packages = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "guppy",
    "certauth.certauth",
    "prctl",
]
ignore_missing_imports = true
